Loaded all boilerplates
Bootstrapper.html is also at 0.4 and ready for some use! It only needs
to pass it's final unit test to get to 0.5 and then full input type
checking, as well as any FailureCases I can think of before reaching 1.0
and ready for general consumption.

The major challenge with this will be getting all the documents into
shape - they aren't bad, but there's a lot of work to do.