Warehouser heading to 0.5
Warehouser nearly 0.5 - just finalizing the format of security that will
be used.

For the first spec it has been determined the best option is for it to
have it as a server issued client id that can configurably be
transferrable by IP or not.

It will also securely allow a statically configured session variable
that is read and can either be truthy/falsy which will be replaced by a
unique id of the current microtime with a random value, or containing a
string username.

Anonymous requests will either be allowed or not by schema in version 2.